fix(ci): keep test files under the server size guard
The new arbiter-consumption regression test pushed
tests/arbiter_park_transaction.rs over the 51200-byte size guard —
moved it to its own integration-test file, reusing the crate's
cfg-gated test_helpers via a test-support self dev-dependency instead
of duplicating fixtures. Collapsed the three per-test
await_holding_lock allows in tool_dispatch_tests.rs into one
module-level allow to bring that file back under the cap too.
