+++ server/crates/djinn-agent/tests/arbiter_park_transaction.rs
@@ -1356,66 +1356,3 @@ async fn planner_escalation_close_releases_source() {
     );
 }
 
-/// An approve decision must mark the arbitration row consumed. A row left
-/// unconsumed after approve wedges the task if it re-enters the second-strike
-/// path (e.g. a merge-conflict reopen): the coordinator treats the stale row
-/// as "arbiter already in flight" every tick and dispatches nothing until the
-/// arbitration deadline (incident lre2, 2026-07-16).
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn arbiter_approve_marks_arbitration_row_consumed() {
-    let db = Database::open_in_memory().expect("open in-memory db");
-    let (project_id, epic_id) = create_project_and_epic(&db).await;
-    let task = create_task(&db, &project_id, &epic_id).await;
-
-    let arb_repo = TaskArbitrationRepository::new(db.clone());
-    let failing_jobs = serde_json::json!([]);
-    let ex = serde_json::json!([]);
-    arb_repo
-        .try_create(CreateArbitrationParams {
-            task_id: &task.id,
-            hold_cycle: 0,
-            deadline_at: None,
-            mirror_head_sha: None,
-            github_head_sha: None,
-            pr_url: None,
-            failing_ci_job_ids: &failing_jobs,
-            dossier: None,
-            directive: None,
-            verification_command: None,
-            excluded_models: &ex,
-        })
-        .await
-        .expect("create arbitration row");
-
-    let ctx = test_agent_context(db.clone());
-    let services: Arc<dyn SupervisorServices> =
-        services_for_agent_context(ctx, CancellationToken::new());
-
-    services
-        .record_arbiter_decision(
-            task.id.clone(),
-            "approve".into(),
-            r#"{"summary": "looks good"}"#.into(),
-        )
-        .await
-        .expect("record_arbiter_decision must succeed");
-
-    let record = arb_repo
-        .get_by_task_and_cycle(&task.id, 0)
-        .await
-        .expect("read arbitration row")
-        .expect("arbitration row exists");
-    assert_eq!(
-        record.state, "consumed",
-        "approve must mark the arbitration row consumed"
-    );
-    assert!(
-        record.consumed_at.is_some(),
-        "consumed_at must be stamped on approve"
-    );
-    assert_eq!(
-        record.directive.as_ref().and_then(|d| d["decision"].as_str()),
-        Some("approve"),
-        "decision must be persisted on the row before consumption"
-    );
-}
